Mathematica

Documentation

packageToArcs :: Package -> [(String, String)]

(a, b) means the arc: a -> b

toMathematicaArcs :: Package -> [String]

Convert a package to a list of directed edges: DirectedEdge

toMathematicaList :: Foldable t => t String -> String

Convert a foldable list of strings into a Mathematica list (as a string)

toMathematicaGraph :: [Package] -> String

Convert a list of Packages into a single Mathematica Graph (source)

toCSV :: [Package] -> String

Convert a list of Packages into a single csv String